requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.2.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
milvus-lite==2.4.11
mpmath==1.3.0
numpy
packaging==24.2
pandas==2.2.3
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
pymilvus==2.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
Werkzeug==3.1.3

# Core FastAPI and Retell
fastapi>=0.100.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
retell-sdk>=4.6.0 # Or your specific version
openai>=1.23.0 # Or your specific version for LLM interaction

# Dependencies for FastAPI, Retell, OpenAI, or Pydantic
aiofiles>=23.0.0
annotated-types>=0.6.0
anyio>=4.0.0
bidict>=0.22.0
blinker>=1.7.0
certifi>=2024.0.0
click>=8.0.0
exceptiongroup>=1.2.0
h11>=0.14.0
httpcore>=1.0.0
httpx>=0.26.0
idna>=3.6
importlib-metadata>=7.0.0 # Required for Python < 3.10 by some packages
# itsdangerous # Usually a Flask dep, check if needed by Retell/FastAPI directly
# Jinja2 # Usually for templating, check if needed
# MarkupSafe # Usually a Jinja2 dep
pydantic>=2.6.0
pydantic_core>=2.16.0
python-engineio>=4.8.0 # Used by Retell
python-multipart>=0.0.9 # For FastAPI file uploads, good to keep if any form data expected
sniffio>=1.3.0
starlette>=0.27.0 # FastAPI dependency
tqdm>=4.66.0 # For progress bars, esp. during model downloads
typing_extensions>=4.9.0 # For Pydantic and type hinting
wsproto>=1.2.0

# RAG specific dependencies
torch
transformers
pymilvus
sentence-transformers

# Optional, but can be useful
# cffi # If any underlying C libraries need it
# cryptography # For any direct crypto needs